Function has been tested with unit tests, however a bug occured because of earlier setup with babel and commonJS
Since jest only works with module.exports and require, this is being used. However, in some others files, we cannot use require, because babel transpiles into ES6. This is has to do with node module commonJS 2015, so probably an option has to be changed. Don't merge this commit to master